pyproject.toml
[tool.poetry]
name = "server-transaction"
version = "0.1.0"
description = ""
authors = ["Coldot <41678750+Coldot@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
pytest = "^8.3.5"
boto3 = "^1.37.16"
botocore = "^1.37.16"
httpx = "^0.28.1"
dotenv = "^0.9.9"
xrpl-py = "^4.1.0"
pymongo = "^4.11.3"
motor = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"